Figured out why redirects weren't working. Fixed it

+++ src/include/drivers/pit.h
@@ -1,7 +1,9 @@
 #ifndef __PIT_H__
 #define __PIT_H__
 
-void sleep(int ticks);
+#include <stdint.h>
+
+void sleep(uint64_t ticks);
 int init_pit();
 
 #endif
